+++ centimeter/utils/editorTypes.ts
@@ -1,0 +1,9 @@
+export interface SlidesDropdown {
+  dropdown: string;
+  typeOfSlide: string;
+}
+
+export interface MakePresentationSideBar {
+  name: string;
+  ownComponents: string[];
+}
