+++ centimeter/utils/userTypes.ts
@@ -1,0 +1,9 @@
+export interface User {
+  id?: string;
+  email?: string;
+  username?: string;
+  first_name?: string;
+  last_name?: string;
+  access?: string;
+  refresh?: string;
+}
